// src/modules/colors.d.ts
export declare function randomColor(text: string): string;
export declare const theme: {
    art: (str: string) => string;
    weather: (str: string) => string;
    divider: () => string;
    word: (str: string) => string;
    category: (str: string) => string;
    definition: (str: string) => string;
};
declare const _default: {
    randomColor: typeof randomColor;
    theme: {
        art: (str: string) => string;
        weather: (str: string) => string;
        divider: () => string;
        word: (str: string) => string;
        category: (str: string) => string;
        definition: (str: string) => string;
    };
};
export default _default;